requirements.txt
accelerate==0.30.1
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
datasets==2.19.2
dill==0.3.8
filelock==3.14.0
frozenlist==1.4.1
fsspec==2024.3.1
huggingface-hub==0.23.2
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
numpy==1.26.4
packaging==24.0
pandas==2.2.2
psutil==5.9.8
pyarrow==16.1.0
pyarrow-hotfix==0.6
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
six==1.16.0
sympy==1.12.1
tokenizers==0.19.1
torch==2.2.2
tqdm==4.66.4
transformers==4.41.2
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
xxhash==3.4.1
yarl==1.9.4
